Enable Flash Attention 2 properly and add fused AdamW
- Fix train_autoreg.py: add torch_dtype=torch.bfloat16 to from_config so FA2
  actually activates (previously fell back to standard attention due to float32 dtype)
- Add --optim adamw_torch_fused for ~5-10% optimizer speedup

+++ Training Scripts/train_autoreg.py
@@ -570,7 +570,8 @@ def main():
     else:
         model = AutoModelForCausalLM.from_config(
             config, trust_remote_code=model_args.trust_remote_code,
-            attn_implementation="flash_attention_2"
+            attn_implementation="flash_attention_2",
+            torch_dtype=torch.bfloat16,
         )
         n_params = sum(
             {p.data_ptr(): p.numel() for p in model.parameters()}.values()
